refactor: reorganize store structure and remove unused interfaces
- Deleted `http-response.interface.ts` and `token-key.type.ts` from `src/types`.
- Deleted `store.tsx` from `src/_services`.
- Created new files for store context, reducer, and initial state in `src/_services/store`.
- Renamed and moved several type definitions to centralize locations.
- Updated imports in various files to reflect new paths and eliminate circular dependencies.

This refactor streamlines the project structure for the store and improves maintainability.
