+++ src/_application/checkAnswer.ts
@@ -1,4 +1,5 @@
 import { MouseEvent } from 'react';
+
 import { QuizStorageService } from './ports.ts';
 
 type Dependencies = {
