+++ src/_application/fetchToken.ts
@@ -1,5 +1,5 @@
 import { quizApiService } from '../_services/quiz-api.service.ts';
-import { localStorage } from '../_services/storageAdapter.ts';
+import { localStorage } from '../_services/store/storageAdapter.ts';
 
 export const fetchToken = async (
     setToken: () => void,
