+++ src/_services/quiz-api.service.ts
@@ -1,10 +1,9 @@
-import { Difficulty } from '../types/difficulty.enum.ts';
-
 import { QuestionRdo, QuizApiService, ResetTokenRdo, TokenRdo } from '../_application/ports.ts';
-import { apiService } from './api.service.ts';
 import { getFetchQuestionsUrl } from '../_lib/getFetchQuestionsUrl.ts';
 import { getFetchTokenUrl } from '../_lib/getFetchTokenUrl.ts';
 import { getResetTokenUrl } from '../_lib/getResetTokenUrl.ts';
+import { Difficulty } from '../types/difficulty.enum.ts';
+import { apiService } from './api.service.ts';
 
 export const quizApiService: QuizApiService = {
     fetchQuestions: async (amount: number, difficulty: Difficulty, token: string) => {
