+++ src/_services/store/storageAdapter.ts
@@ -1,6 +1,6 @@
-import { LocalStorageService, QuizStorageService } from '../_application/ports.ts';
-import { useStore } from './store.tsx';
+import { LocalStorageService, QuizStorageService } from '../../_application/ports.ts';
 import { localStore } from './local-store.ts';
+import { useStore } from './useStore.ts';
 
 export function useQuizStorage(): QuizStorageService {
     return useStore();
