+++ src/_services/useCheckAnswer.ts
@@ -1,6 +1,7 @@
-import { useQuizStorage } from './storageAdapter.ts';
-import { checkAnswer } from '../_application/checkAnswer.ts';
 import { MouseEvent } from 'react';
+
+import { checkAnswer } from '../_application/checkAnswer.ts';
+import { useQuizStorage } from './store/storageAdapter.ts';
 
 export const useCheckAnswer = () => {
     const quizStorage = useQuizStorage();
