+++ src/_services/useNextQuestion.ts
@@ -1,5 +1,5 @@
-import { useQuizStorage } from './storageAdapter.ts';
 import { nextQuestion } from '../_application/nextQuestion.ts';
+import { useQuizStorage } from './store/storageAdapter.ts';
 
 export const useNextQuestion = () => {
     const quizStorage = useQuizStorage();
