+++ src/_services/useStartTrivia.ts
@@ -1,5 +1,5 @@
-import { useQuizStorage } from './storageAdapter.ts';
 import { startTrivia } from '../_application/startTrivia.ts';
+import { useQuizStorage } from './store/storageAdapter.ts';
 
 export const useStartTrivia = () => {
     const quizStorage = useQuizStorage();
