+++ src/main.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
+
+import { Provider } from './_services/store/store.tsx';
 import App from './_ui/App.tsx';
-import { Provider } from './_services/store.tsx';
 
 const ROOT_ELEMENT_ID = 'root';
 const rootElement = document.getElementById(ROOT_ELEMENT_ID);
