+++ src/shared-kernel.d.ts
@@ -1,0 +1,6 @@
+interface HttpResponse<T> {
+    status: number;
+    data: T;
+}
+
+type TokenKey = 'sessionToken';
