--- src/types/http-response.interface.ts
@@ -1,4 +1,0 @@
-export interface HttpResponse<T> {
-    status: number;
-    data: T;
-}
