+++ src/types/initial-state.type.ts
@@ -1,5 +1,5 @@
+import { Answer } from '../_domain/answer.type.ts';
 import { QuestionsState } from './question-state.type.ts';
-import { Answer } from '../../_domain/answer.type.ts';
 
 export type InitialState = {
     loading: boolean;
