+++ src/types/question-state.type.ts
@@ -1,3 +1,3 @@
-import { Question } from '../../_domain/question.type.ts';
+import { Question } from '../_domain/question.type.ts';
 
 export type QuestionsState = Question & { answers: string[] };
