--- src/types/token-key.type.ts
@@ -1,1 +1,0 @@
-export type TokenKey = 'sessionToken';
